+++ torch_frame/utils/__init__.py
@@ -1,6 +1,8 @@
-from .io import save_tf, load_tf
+from .io import save, load
+from .concat import cat
 
 __all__ = functions = [
-    'save_tf',
-    'load_tf',
+    'save',
+    'load',
+    'cat',
 ]